pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-to-notebooklm"
version = "1.0.0"
description = "将 GitHub 项目自动导入 NotebookLM 进行 AI 分析"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "OpenClaw", email = "openclaw@example.com"}
]
keywords = ["github", "notebooklm", "ai", "code-analysis", "openclaw"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "gitingest>=0.1.4",
    "notebooklm-py>=0.1.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
browser = [
    "notebooklm-py[browser]",
    "playwright>=1.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
github-to-notebooklm = "github_to_notebooklm.core:main"
g2n = "github_to_notebooklm.core:main"

[project.urls]
Homepage = "https://github.com/openclaw/github-to-notebooklm"
Repository = "https://github.com/openclaw/github-to-notebooklm"
Issues = "https://github.com/openclaw/github-to-notebooklm/issues"

[tool.setuptools]
packages = ["github_to_notebooklm"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
